Move CommonCode.h out of build_all directory
Move the file closer to the rest of the C++ code; within the
`_LowLevelCode/cpp` directory.
