+++ _LowLevelCode/cpp/accumulate_cut_c/accumulate_cut_c.h
@@ -1,7 +1,7 @@
 #ifndef H_ACCUMULATE_CUT
 #define H_ACCUMULATE_CUT
 
-#include "../../build_all/CommonCode.h"
+#include "../CommonCode.h"
 
 #define iRound(x) (int)floor((x) + 0.5)
 //
