+++ _LowLevelCode/cpp/bin_pixels_c/bin_pixels_c.cpp
@@ -1,4 +1,4 @@
-#include "../../build_all/CommonCode.h"
+#include "../CommonCode.h"
 //
 enum input_arguments {
     Sqw_parameters,
