+++ _LowLevelCode/cpp/mtimesx_horace/CMakeLists.txt
@@ -5,7 +5,7 @@ set(
 
 set(
     HDR_FILES
-    "../../build_all/CommonCode.h"
+    "../CommonCode.h"
     "MatMultiply.h"
 )
