+++ _LowLevelCode/cpp/sort_pixels_by_bins/CMakeLists.txt
@@ -5,7 +5,7 @@ set(
 
 set(
     HDR_FILES
-    "../../build_all/CommonCode.h"
+    "../CommonCode.h"
     "sort_pixels_by_bins.h"
 )
